discord-bot/modules/stubs.py
import discord
from discord import ui
import time
import random
from data_manager import dm
from logger import logger

class AntiRaidSystem:
    def __init__(self, bot):
        self.bot = bot

    async def start_monitoring(self):
        pass

    async def handle_member_join(self, member):
        pass

    async def handle_member_remove(self, member):
        pass

    async def handle_message(self, message):
        pass

class GuardianSystem:
    def __init__(self, bot):
        self.bot = bot

    async def start_monitoring(self):
        pass

    async def handle_message(self, message):
        pass

class AutoModSystem:
    def __init__(self, bot):
        self.bot = bot

    async def handle_message(self, message):
        pass

class WarningsSystem:
    def __init__(self, bot):
        self.bot = bot

class StaffPromotionSystem:
    def __init__(self, bot):
        self.bot = bot

class StaffShiftSystem:
    def __init__(self, bot):
        self.bot = bot

    async def handle_message(self, message):
        pass

    async def start_tasks(self):
        pass

class StaffReviewSystem:
    def __init__(self, bot):
        self.bot = bot

    async def start_monitoring(self):
        pass

class StarboardSystem:
    def __init__(self, bot):
        self.bot = bot

    async def handle_reaction_add(self, reaction, user):
        pass

class AIChatSystem:
    def __init__(self, bot):
        self.bot = bot

    async def handle_message(self, message):
        pass

class ApplicationSystem:
    def __init__(self, bot):
        self.bot = bot

    def get_persistent_views(self):
        return []

class AppealSystem:
    def __init__(self, bot):
        self.bot = bot

    def get_persistent_views(self):
        return []

class ModmailSystem:
    def __init__(self, bot):
        self.bot = bot

    async def handle_dm(self, message):
        pass

    def get_persistent_views(self):
        return []

class AnnouncementSystem:
    def __init__(self, bot):
        self.bot = bot

    async def start_monitoring(self):
        pass

class AutoResponderSystem:
    def __init__(self, bot):
        self.bot = bot

    async def handle_message(self, message):
        pass

class ReactionRoleSystem:
    def __init__(self, bot):
        self.bot = bot

class ReactionMenuSystem:
    def __init__(self, bot):
        self.bot = bot

class RoleButtonSystem:
    def __init__(self, bot):
        self.bot = bot

class ModerationSystem:
    def __init__(self, bot):
        self.bot = bot

class LoggingSystem:
    def __init__(self, bot):
        self.bot = bot

class ModLoggingSystem:
    def __init__(self, bot):
        self.bot = bot