discord-bot/requirements.txt
discord.py>=2.3.0
python-dotenv>=1.0.0
cryptography>=41.0.0
aiosqlite>=0.19.0
tenacity>=8.2.0
pytz>=2023.3